Move soft keyword definitions outside the parser
